Move getSizeInfo to _staging
JIRA: RAIL-4170
