+++ libs/sdk-ui-dashboard/src/_staging/dashboard/dashboardLayout.ts
@@ -19,7 +19,7 @@ import keyBy from "lodash/keyBy";
 import invariant, { InvariantError } from "ts-invariant";
 import { validateDashboardLayoutWidgetSize } from "../../presentation/layout/DefaultDashboardLayoutRenderer/utils/sizing";
 import { ObjRefMap } from "../metadata/objRefMap";
-import { MeasurableWidgetContent } from "../../model/layout/sizing";
+import { MeasurableWidgetContent } from "../../_staging/layout/getSizeInfo";
 
 function extractContentFromWidget(
     widget: IDashboardWidget,
