+++ libs/sdk-ui-dashboard/src/presentation/widget/kpiPlaceholder/KpiPlaceholderConfigurationPanel.tsx
@@ -15,7 +15,7 @@ import {
     useDashboardDispatch,
     useDashboardSelector,
 } from "../../../model";
-import { getSizeInfo } from "../../../model/layout";
+import { getSizeInfo } from "../../../_staging/layout/getSizeInfo";
 
 import { KpiConfigurationPanelCore } from "../kpi/EditableDashboardKpi/KpiConfigurationPanel/KpiConfigurationPanelCore";
 import { KpiPlaceholderWidget } from "../../../widgets/placeholders/types";
